[UI/Engine] Add an InfoBar-like UI to register/ignore a work
This commit provides a baseline of InfoBar to insert pieces of UI into the host
page. This UI (intrusive for the moment) can be used for various purposes,
including interaction with the user and notification of background work.

To do so, this commit adds the following pieces:
 - Engine bits to provide the registration/aliasing/ignore logics, and make the
   link with the storage component.
 - UI bits to spawn an iframe-style InfoBar, with a parameterized HTML content
 - register-diag page (html + scripts/) for the comic registration dialog
   within the InfoBar iframe
 - Communication (not working ATM) between the register-diag & the origin
   window to allow closing the iframe after the user has interacted with it

Note that a few pieces are still not working, such as the messaging channel
between the host page and the inserted iframe content (register-diag.html +
scripts/register-diag.js)