[UX] Make SideBar and InfoBar interact properly
In order to reach the desired UX, we needed the following:
 - Have the InfoBar get hidden when the SideBar is Shown
 - Have the InBar get spawned again when the SideBar is Hidden

To do so, a few steps were taken:
 - Always start by spawning the SideBar
 - Setup the event code which will spawn the InfoBar when Hiding the
   SideBar
 - Setup the event code which will show the SideBar when clicking on the
   '+' button of the InfoBar
 - Setup the code which will hide the InfoBar when Showing the SideBar
   (which is thus automatically called when we click the '+' button of
   the Infobar -> Only one code path for both showing the sidebar and
   clicking the '+' button)
 - For overlapping correctness, z-indexes of the two bars were inverted

As for the details, as both bars are iframe generated by the extension's
domain, it is possible to run code in one iframe from the other. This is
the technique used to communicate between these two iframes here in this
commit.

Additionally, these changes lead to having the SideBar act as the
main/core controller of the addon, while the Infobar is purely there as
a UX element, rather than a functional necessity.