[Sidebar] Add  a `+` button to record a new manga
This commit adds a register button (`+`) in the sidebar.

The register button is located next to the searchbox input area, and is only
shown when the sidebar is used in `REGISTER` mode (as opposed to `BROWSE`
mode).

When a user click on the `Register` button (`+`) from the infobar, the sidebar
is shown with its own register button shown, so that the user may either click
on an already tracked manga (this creates an alias to the current source for
that manga), or fill the searchbox and click on the register button (`+`) to
add a new tracked manga, based on the current reader & manga displayed on the
page.