[Sidebar] Add a color transition effect to provide visual feedback
This commit adds a color transition effect on the toggle button to provide a
visual feedback to the user about the success or failure of the last operation
(Registering a comic, Tracking, Aliasing a comic, removing a comic).

This transition is controlled through the Javascript code of the sidebar, and
goes from red to the default color for a failure, while it goes from green to
the default color for a success.

The transition is for now triggered in the following situations:
 - After a click on the sidebar's register button
 - After a click on an comic to create an alias
 - After updating a registered comic's tracking automatically on load

Note that to support the tracking notification, the cross-frame messaging
system is used.