[Engine/DataAPI] Fix storage usage: always access object's key after get
We were previously having mistaken expectations for the storage API from the
browser, expecting to only get the payload for one given key.

The reality is slightly different: All requests get a response with an object,
where the requested key may (if some data or default value is available) be
present, and index the actual payload for that key in the storage.

This commit updates DataAPI to use both the default-value mechanism of storage
as well as properly dereference any retrieved data using the requested key, to
ensure we always manipulate the payload only.