[Engine/DataModel] Make ComicSource serialization deterministic (key order)
The serialization of the BmcComicSource class was not deterministic previously,
and the order of insertion could have led to a mistmatch of data that should
have matched.

This commit provides the guarantee, using the right method with JSON.stringify,
that the fields will always be ordered in the serialized payload.

Fixes issue introduced by commit 69a763c39c