[Engine] Allow back-and-forth messaging between host window and webext frames
In order to ensure that the messaging system would work when the host window is
sending messages to one of our iframes, it was necessary to accept messages
from the host window's origin.

To do so, the origin is transmitted from the entrypoint script through the
engine, and to the sidebar (via an additional query string parameter).

This allows proper two-way communication between the host window and any
injected iframe (sidebar here).