[Engine/Memoization] Allow Engine to reset its Comic ID Memoization
This commit fixes the situation from issue #13 through two changes:
 1. Add an internal method to reset the memoization
 2. Call the memoization reset method after Successful
    registration/alias

With this, any usage of the Engine's utility methods will account for
the update of the memoization triggered by the alias/register wrappers.

Note that while this is an improvement, it does not forcefully trigger
the memoization reset of any BmcEngine instance not directly used in the
registration/alias process.

Fixes #13