[Refactor] Decouple BmcEngine from Sidebar
In order to better identify the role of each module within the
web-extension, it was necessary to ensure that only one instance of each
logic-critical unit was existing at any time on the whole page.

To do so, the sidebar scripts are now using messages to request Data
operations to the Core (BmcEngine), and using utility classes
(BmcDataAPI and BmcMessagingHandler) to handle minor tasks.

As such, this commit:
 - Adds a message handling in BmcEngine, to act as messaging proxy for
   methods such as register & alias
 - Use the BmcEngine mesages from the sidebar to instruct the core to
   register/alias a comic
 - Adds a message handling in the Sidebar to allow toggling it from any
   other frame
 - Use the toggle message from the UI module to show/hide the sidebar