[UI/sidebar-toggle] Use the new utility function
The UI code was previously attemptint to use a function from the sidebar
scripts directly. This might have been due to the sidebar script calling
this code somehow, or was a bug lying in wait.

This commit fixes this by using the messaging system hidden away by the
`toggleSidebar()` utility function.