[Sidebar] Align roles of the two registration buttons
This commit re-aligns the roles of the button-add and register-but DOM
elements for the aim of removing the Infobar to the profit of the
sidebar.

The new 'button-add' (to be renamed) is now the proper trigger of the
registration of a comic into the DB; while the former 'register button'
now forces open the sideBar into the Register mode.

This commit also provides a slight refactor of the code moving the
buttons around in case of toggling the SideBar.

Work towards completion of #12