[Design/Background] Handle All supported sites in a generic manner
In order to allow computing informations for various supported readers
seamlessly, from any piece of content_script, it was necessary to have a piece
of javascript loaded in a privileged context.

The easiest and most standard way was to use a background script, which loads a
number of scripts, and handles common requests for all content_scripts injected
into the various relevant tabs. This is also a potentially good way to cache
something commonly to all instances of the add-on, throughout the available tabs.

To do so, this commit provides the following changes:

- Extract `getBrowser()` function into a common utils.js file

- Harmonize API of all support/*.js scripts, so that only one Class model is
  used (through duck-typing rather than inheritance)

- Add a source.js with a BmcSources class which will handle resolving the calls
  to the right reader instance, according to the origin (or reader URL) provided

- Add communication channel in messaging.js for background-script <->
  content_scripts exchanges. Thanks to this, the content_scripts will be able
  to request common information to the background script.

- Add a background script, which will handle computations on the various
  Reader's URLs (parsing, generating)
  - Update manifest.json to load all sources scripts as dependency of
    background-engine.js

- Rework entrypoint.js to request URL parsing from the background script, and
  only then instanciate the Engine (and thus UI).