[DataModel] Use partially-opaque JS object for URL data
Each supported reader might have its own specificities. As an example, some
readers offer multiple languages to browse (and/or) read comics and mangas.
This should be included as part of the information retrieved from the URL (is
possible), so that we can later rebuild that exact URL.

In order to be the least intrusive possible, we're bundling the custom data in
the "comicInfo" object provided by the parseURL functions, including a "common"
field, defined by BookMyComics's internal APIs (and thus includes name, chapter
and page). Additional custom fields can be added at will, and will be kept in
the storage.

This change provides good progress towards solving #14, and allowing to browse
the tracked works from the SideBar.