[Sidebar] Fix search to only filter based on Comic Label
When the code was modified to list the tracked commits from the storage, we
forgot to update the searching logic, which seemed to work at the time.
As it relies on the DOM's element's innerText, it was actually also including
the string from the Comic's label, which explains why it seemed to work.

Alas, since that innerText also included the reader names (for the alias
listing), searching for a string which matches one reader's name also ended up
selecting all comics wich were referencing that reader, screwing the expected
results.

This commit fixes the situation by ensuring that we only attempt to match the
searchbox's content with the actual label for the Comics.

Fixes #7, introduced by commit ca1bf083