[Engine] Fix storage wrapper promise/callback handling
It seems like some versions of chrome don't even accept access to undefined
variables such as "browser". This made the current check fail with a thrown
Error to start with.

Then, in the browsers (Chrome, Firefox) APIs, no storage function is expected
to provide an error value to the callbacks, only results. As such, the provided
wrapper to handle both callbacks and promises was handling parameters in an
inverted order, compared to the expectations from the BmcStorage class's API.

As such, this commit:
 - Fixes the initialization of the Storage module by catching ReferenceError
   instances, and keeping the callback mode in this case. A side comment
   additionally explains this.
 - Circumvent the parameter inversion in the case of callback mode by inverting
   the order in the wrapper function, so that the BmcStorage module's callback
   API stays stable. Note that as no error is supposed to be provided, a piece
   of the code seems like dead-code, but was kept for compatibility in regards
   to promise handling.

Fixup storage again

Fixup storage detection