[SideBar] Add delete icon to Comic & Sources in MangaList
This commit aims to provide the UI groundwork to allow deleting Source and
Comics from the Manga Listing.

The delete action is represented by a trashbin icon, which only appears when
hovering on the correponding line (Either the comic label or the source's
label).

The following changes only provides the UI change, and not the actual operation
implementation (only logging for validation).

Addresses part of #26