[DataModel] Prevent registration of an already existing label
This commit adds a new mapping Label/ID in the datamodel, so that we can use it
to check for existence of a label.

Thanks to this, the web-extension now guarantees that user-specified labels are
unique, and will show an error upon re-using an existing label for a new comic
registration.