[UI/Tests] Include the BMC iframe in the body instead of after it
Due to focus-related issue of remote webdriver control, the tests often failed
to re-focus into the BMC iframe to inspect or control it.

Indeed, the WebDriver reported the frame not to be part of the currently
focused element (the body, by default), and thus refused to switch focus to the
BMC frame.

Moving the frame into the body, rather than inserting it as the last child of
the `html` tag greatly improved the stability of the tests, and seems more
aligned to the HTML expectations (content in the body).