[Tests/Readers] Enhance reliability of load_random()
Some of the incoming tests will rely on `load_random()` to successively find
comics to browse.
Alas, there is a risk of stumbling twice on the same comic despite the test
attempting to find two different ones.

To solve this issue, the readers are now able to resolve a "comic name" from
the current URL (which can be used by the testing code), and the
`load_random()` function is now able to take a predicate which will be executed
using the reader_driver as a parameter.

The addition of these two changes will allow new tests to guarantee that a
randomly loaded comic is a different one that a previous one.