[Tests] Work-around for #50
In order to ensure that the current tests succeed despite issue #50 (sidebar
not properly reloading its data after registration/deletion), this commits adds
the explicit refresh() of the pages, along with a new capacity of the
`BmcController`, which is to reset its internal SideBar state.

This new capacity will be useful beyond the work-around for #50, as it helps
ensuring that no stale state/WebElement objects are kept by the controller
after a change of URL. Such stale objects would otherwise break the Selenium
APIs, and thus the tests.