[SidePanel/Adder] Improve error notification on registration
This commit adds a small element in the Adder mode of the SidePanel, which can
display a short error message to the user.

This element is now updated through various processes:
 - Updated upon any update of the label input field, the availability of the
   label in the storage is checked, triggerring an early error notification if
   need be
 - Updated upon notification of a registration event
 - Cleaned upon switching from any other panel to the Adder mode of the
   SidePanel

When this new element is used for displaying an error, the registration
confirmation button is disabled for consistency.

The tests (TestRegister::test_eexist) have been updated accordingly with a new
utility to check for registration errors through the inspection of the new
element.