[Support] Replace manifests '*' protocol by http for mangahere
The mentionned website seems to no longer have valid certificates for their SSL
connection, and as such, our tests are now failing when relying on HTTPS
instead of HTTP.

This commit updates the manifests to ensure that we no longer fall on the
insecure certificate message during our testing. Incidentally, it also limits
the use of the web-extension to the non-https version, but we deem it
acceptable as long as the certificates are not trusted by default.

Relates to #72