[Tests/Utils] Rework reader.load_random() predicate handling
This commit reworks the load_random() predicate handling to unify it into a
simple behavior.

The new predicate is handled by a decorator which raises an exception if the
predicate is not satisfied, while eating-up the `predicate` argument.
This should allow for more flexibility in the use of this predicate decorator,
while we can use it to wrap all `load_random()` methods without breaking its
actual API. The parametrable error type will also make this decorator
compatible with the `retry` decorator which depends on `RetriableError`.