#23 [Engine/Tracking] Facilitate AJAX handling by moving tracking setup in Engine
The Engine is now able to detect changes to the URL, and trigger a tracking
operation upon this event. As this changes the dependency between some of the
javascript files, they have been reordered in the manifests.

Additionally, this revealed that the engine was unused in the sidebar's code,
which is now cleaned-up.