[CI] Update strings checker tool & code to fix erroneous errors
The tool now properly accounts for the errors in the string's usage count,
preventing erroneous 0-counts when the only the error is used in the error
mapping, which gave the false impression that it was never used, while only the
error code was actually never referenced.

Additionally, as the tool is unable to detect strings usage in all
indentations/format styles, the code is updated to fit the tool detection
capabilities.