[Options/Storage] Add auto-migration of data upon storage engine update
In order to persist the settings, so that we do not loose track of the
data, a new `BmcSettings` object is introduced to handle the persistence
of our settings in the browser's local storage.

The usage of BmcStorageFactory was thus updated to ensure consitency of
behavior, and this, in turn, led to a need to rework parts of existing
javascript code as we needed to ensure that settings were properly
loaded from local storage before instanciating the BmcDataAPI that
indirectly depended on it.

Additionally, when swapping storage engines, the auto-migration properly
cleans up any old data of the source storage, once the migration is
complete, and the settings updated.